requirements.txt
numpy>=1.23.5
pandas>=1.5.3
matplotlib>=3.7.1
seaborn>=0.12.2
scikit-learn>=1.2.2
scikit-surprise>=1.1.3
streamlit>=1.23.1
plotly>=5.15.0
